+++ histore/archive/manager/base.py
@@ -23,7 +23,7 @@ class ArchiveManager(metaclass=ABCMeta):  # pragma: no cover
 
         Returns
         -------
-        dict
+        dict(string: histore.archive.manager.descriptor.ArchiveDescriptor)
         """
         raise NotImplementedError()
 
@@ -93,13 +93,31 @@ class ArchiveManager(metaclass=ABCMeta):  # pragma: no cover
 
         Returns
         -------
-        histore.archive.vase.Archive
+        histore.archive.base.Archive
 
         Raises
         ------
         ValueError
         """
         raise NotImplementedError()
+
+    def get_by_name(self, name):
+        """Get descriptor for the archive with the given name. If no archive
+        with that name exists None is returned.
+
+        Parameters
+        ----------
+        name: string
+            Archive name
+
+        Returns
+        -------
+        histore.archive.manager.descriptor.ArchiveDescriptor
+        """
+        for archive in self.archives().values():
+            if archive.name() == name:
+                return archive
+        return None
 
     def list(self):
         """Get the list of descriptors for the maintained archives.
@@ -109,3 +127,21 @@ class ArchiveManager(metaclass=ABCMeta):  # pragma: no cover
         list(histore.archive.manager.descriptor.ArchiveDescriptor)
         """
         return list(self.archives().values())
+
+    @abstractmethod
+    def rename(self, identifier, name):
+        """Rename the specified archive. Raises a ValueError if the identifier
+        is unknown or if an archive with the given name exist.
+
+        Parameters
+        ----------
+        identifier: string
+            Unique archive identifier
+        name: string
+            New archive name.
+
+        Raises
+        ------
+        ValueError
+        """
+        raise NotImplementedError()