+++ histore/archive/manager/fs.py
@@ -23,8 +23,8 @@ class PersistentArchiveManager(ArchiveManager):
     """The persistent archive manager maintains a set of archives on disk. The
     list of archive descriptors is also maintained on disk as a Json file.
     """
-    def __init__(self, basedir=None):
-        """Initialize the base directory under which all archives are stores in
+    def __init__(self, basedir=None, exists=False):
+        """Initialize the base directory under which all archives are stored in
         individual sub-folders. If the base directory is not given the value
         will be read from the environment variable HISTORE_BASEDIR or the
         default value $HOME/.histore.
@@ -35,15 +35,26 @@ class PersistentArchiveManager(ArchiveManager):
         ----------
         basedir: string
             Path to dorectory on disk where archives are maintained.
+        exists: bool, default=False
+            Raise an error if the given base directory is not empty or contains
+            the descriptor file.
         """
         if basedir is None:
             basedir = config.BASEDIR()
-        self.basedir = util.createdir(basedir)
         # Initialize path to file that maintains archive descriptors.
-        self.descriptorfile = os.path.join(self.basedir, 'archives.json')
+        self.descriptorfile = os.path.join(basedir, 'archives.json')
+        exists_descriptorfile = os.path.isfile(self.descriptorfile)
+        # Raise error if the archive manager base directory is expected to
+        # exist.
+        if exists:
+            # Raise error if (i) the directory does not exists, or (ii) exists
+            # but does not contain the descriptor file.
+            if not os.path.isdir(basedir) or not exists_descriptorfile:
+                raise ValueError('archive manager does not exist')
+        self.basedir = util.createdir(basedir)
         # Initialize the internal cache of archive descriptors
         self._archives = dict()
-        if os.path.isfile(self.descriptorfile):
+        if exists_descriptorfile:
             with open(self.descriptorfile, 'r') as f:
                 doc = json.load(f)
             for obj in doc:
@@ -56,7 +67,7 @@ class PersistentArchiveManager(ArchiveManager):
 
         Returns
         -------
-        dict
+        dict(string: histore.archive.manager.descriptor.ArchiveDescriptor)
         """
         return self._archives
 
@@ -64,7 +75,8 @@ class PersistentArchiveManager(ArchiveManager):
         self, name=None, description=None, primary_key=None, encoder=None,
         decoder=None
     ):
-        """Create a new archive object.
+        """Create a new archive object. Raises a ValueError if an archive with
+        the given name exists.
 
         Parameters
         ----------
@@ -84,7 +96,14 @@ class PersistentArchiveManager(ArchiveManager):
         Returns
         -------
         histore.archive.manager.descriptor.ArchiveDescriptor
+
+        Raises
+        ------
+        ValueError
         """
+        # Ensure that the archive name is unique.
+        if self.get_by_name(name) is not None:
+            raise ValueError("archive '{}' already exists".format(name))
         # Create the descriptor for the new archive.
         descriptor = ArchiveDescriptor.create(
             name=name,
@@ -97,8 +116,7 @@ class PersistentArchiveManager(ArchiveManager):
         primary_key = descriptor.primary_key()
         # Write list of archive descriptors.
         self._archives[identifier] = descriptor
-        with open(self.descriptorfile, 'w') as f:
-            json.dump([d.doc for d in self._archives.values()], f)
+        self.write()
         return descriptor
 
     def delete(self, identifier):
@@ -110,10 +128,11 @@ class PersistentArchiveManager(ArchiveManager):
             Unique archive identifier
         """
         if self.contains(identifier):
-            shutil.rmtree(os.path.join(self.basedir, identifier))
+            archdir = os.path.join(self.basedir, identifier)
+            if os.path.isdir(archdir):
+                shutil.rmtree(archdir)
             del self._archives[identifier]
-            with open(self.descriptorfile, 'w') as f:
-                json.dump([d.doc for d in self._archives.values()], f)
+            self.write()
 
     def get(self, identifier):
         """Get the archive that is associated with the given identifier. Raises
@@ -126,7 +145,7 @@ class PersistentArchiveManager(ArchiveManager):
 
         Returns
         -------
-        histore.archive.vase.Archive
+        histore.archive.base.Archive
 
         Raises
         ------
@@ -154,3 +173,37 @@ class PersistentArchiveManager(ArchiveManager):
             encoder=encoder,
             decoder=decoder
         )
+
+    def rename(self, identifier, name):
+        """Rename the specified archive. Raises a ValueError if the identifier
+        is unknown or if an archive with the given name exist.
+
+        Parameters
+        ----------
+        identifier: string
+            Unique archive identifier
+        name: string
+            New archive name.
+
+        Raises
+        ------
+        ValueError
+        """
+        archive = self._archives.get(identifier)
+        if archive is None:
+            raise ValueError("unknown archive '{}''".format(identifier))
+        if archive.name() == name:
+            # Do nothing if the archive name matches the new name.
+            return
+        # Raise an error if another archive with the new name exists.
+        # Ensure that the archive name is unique.
+        if self.get_by_name(name) is not None:
+            raise ValueError("archive '{}' already exists".format(name))
+        archive.rename(name)
+        # Write list of archive descriptors.
+        self.write()
+
+    def write(self):
+        """Write the current descriptor set to file."""
+        with open(self.descriptorfile, 'w') as f:
+            json.dump([d.doc for d in self._archives.values()], f)