+++ histore/archive/manager/mem.py
@@ -28,7 +28,7 @@ class VolatileArchiveManager(ArchiveManager):
 
         Returns
         -------
-        dict
+        dict(string: histore.archive.manager.descriptor.ArchiveDescriptor)
         """
         return self._descriptors
 
@@ -36,7 +36,8 @@ class VolatileArchiveManager(ArchiveManager):
         self, name=None, description=None, primary_key=None, encoder=None,
         decoder=None
     ):
-        """Create a new volatiole archive object.
+        """Create a new volatile archive object. Raises a ValueError if an
+        archive with the given name exists.
 
         Parameters
         ----------
@@ -57,7 +58,14 @@ class VolatileArchiveManager(ArchiveManager):
         Returns
         -------
         histore.archive.manager.descriptor.ArchiveDescriptor
+
+        Raises
+        ------
+        ValueError
         """
+        # Ensure that the archive name is unique.
+        if self.get_by_name(name) is not None:
+            raise ValueError("archive '{}' already exists".format(name))
         # Create the descriptor for the new archive.
         descriptor = ArchiveDescriptor.create(
             name=name,
@@ -93,12 +101,39 @@ class VolatileArchiveManager(ArchiveManager):
 
         Returns
         -------
-        histore.archive.vase.Archive
+        histore.archive.base.Archive
 
         Raises
         ------
         ValueError
         """
         if identifier not in self._archives:
-            raise ValueError('unknown archive {}'.format(identifier))
+            raise ValueError("unknown archive '{}''".format(identifier))
         return self._archives[identifier]
+
+    def rename(self, identifier, name):
+        """Rename the specified archive. Raises a ValueError if the identifier
+        is unknown or if an archive with the given name exist.
+
+        Parameters
+        ----------
+        identifier: string
+            Unique archive identifier
+        name: string
+            New archive name.
+
+        Raises
+        ------
+        ValueError
+        """
+        archive = self._descriptors.get(identifier)
+        if archive is None:
+            raise ValueError("unknown archive '{}''".format(identifier))
+        if archive.name() == name:
+            # Do nothing if the archive name matches the new name.
+            return
+        # Raise an error if another archive with the new name exists.
+        # Ensure that the archive name is unique.
+        if self.get_by_name(name) is not None:
+            raise ValueError("archive '{}' already exists".format(name))
+        archive.rename(name)