+++ tests/archive/manager/test_archive_descriptor.py
@@ -13,12 +13,15 @@ from jsonschema import ValidationError
 
 from histore.archive.manager.descriptor import ArchiveDescriptor
 
+import histore.util as util
+
 
 def test_archive_descriptor():
     """Test methods for creating archive descriptors."""
     # Create descriptor using a dictionary.
     doc = {
         'id': '0000',
+        'createdAt': util.utc_now().isoformat(),
         'name': 'My Archive',
         'description': 'This is my archive',
         'primaryKey': ['SSN'],
@@ -32,7 +35,7 @@ def test_archive_descriptor():
     assert descriptor.primary_key() == ['SSN']
     assert descriptor.encoder() == 'myencoder'
     assert descriptor.decoder() == 'mydecoder'
-    doc = {'id': '0001'}
+    doc = {'id': '0001', 'createdAt': util.utc_now().isoformat()}
     descriptor = ArchiveDescriptor(doc)
     assert descriptor.identifier() == '0001'
     assert descriptor.name() == '0001'